# docs/openapi.yaml
openapi: 3.1.0
info:
  title: notifyd API
  version: "1.0"
  description: Multi-channel notification delivery API by Fluxintek.
servers:
  - url: https://api.notifyd.fluxintek.com
security:
  - bearerAuth: []
paths:
...
  /notifications/{notificationID}/attempts:
    parameters:
      - name: notificationID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: List delivery attempts
      description: Returns all delivery attempt records for a notification, ordered by attempt number.
      operationId: listAttempts
      tags:
        - Notifications
      responses:
        "200":
          description: List of delivery attempts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeliveryAttempt"
        "400":
          description: Invalid notification ID format.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Notification not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceHmac:
      type: apiKey
      in: header
      name: X-Service-Signature
      description: 'hex(HMAC-SHA256(METHOD + "\n" + request-target + "\n" + X-Service-Timestamp + "\n" + hex(sha256(body)), SERVICE_HMAC_SECRET)). Requires header X-Service-Timestamp (RFC3339), validated within a 5-minute skew window.'
  schemas:
...
    AttemptStatus:
      type: string
      enum:
        - success
        - failure
    DeliveryAttempt:
      type: object
      properties:
        id:
          type: string
          format: uuid
        notification_id:
          type: string
          format: uuid
        attempt_number:
          type: integer
        status:
          $ref: "#/components/schemas/AttemptStatus"
        provider_response:
          type: object
          description: Raw provider API response body.
        error_message:
          type: ["string", "null"]
        duration_ms:
          type: integer
          description: Round-trip duration in milliseconds.
        attempted_at:
          type: string
          format: date-time

    # -------------------------------------------------------------------------
    # Delivery Metrics
    # -------------------------------------------------------------------------
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code or human-readable message.
...